ex03 - adicionando validações

+++ ex3/controller/interfaces/IBaseController.ts
@@ -1,40 +1,27 @@
+import {
+  fail,
+  httpBodyRequest,
+  httpResponse,
+  ok,
+  httpParamsRequest,
+} from "../../http-protocols/http";
 import { IBaseService } from "../../service/interface/IBaseService";
-
-export type httpParamsRequest = {
-  params: string;
-};
-export type httpBodyRequest<T> = {
-  body: T;
-};
-export type httpRequest<T> = httpParamsRequest & httpBodyRequest<T>;
-
-export type httpResponse<T> = {
-  statusCode: number;
-  body: T;
-};
-
-export const ok = <T>(data: T): httpResponse<T> => ({
-  statusCode: 200,
-  body: data,
-});
-
-export type httpResponseError = {
-  statusCode: number;
-  body: any;
-};
-export const fail = (data: any): httpResponseError => ({
-  statusCode: 500,
-  body: data,
-});
-
+import { IValidate } from "../../utils/validate";
 export interface hasId {
   id: string;
 }
 export abstract class IBaseController<T extends hasId> {
-  constructor(protected baseService: IBaseService<T>) {}
+  constructor(
+    protected baseService: IBaseService<T>,
+    protected validateParams: IValidate
+  ) {}
   async save(httpRequest: httpBodyRequest<T>): Promise<httpResponse<T>> {
     try {
       const body = httpRequest.body;
+      const hasCorrectParams = this.validateParams.validate(httpRequest);
+      if (!hasCorrectParams) {
+        throw new Error("missing arguments");
+      }
 
       const response = await this.baseService.save(body);
       return ok(response);
@@ -45,6 +32,9 @@ export abstract class IBaseController<T extends hasId> {
   async edit(httpRequest: httpBodyRequest<T>): Promise<httpResponse<T>> {
     try {
       const id = httpRequest.body;
+      if (!this.validateParams.validate(httpRequest)) {
+        throw new Error("missing arguments");
+      }
 
       const result = await this.baseService.edit(id);
       return ok(result);
@@ -73,6 +63,9 @@ export abstract class IBaseController<T extends hasId> {
   ): Promise<httpResponse<{ message: string }>> {
     try {
       const id = httpParamsRequest.params;
+      if (this.validateParams.validate(httpParamsRequest)) {
+        throw new Error("missing params");
+      }
       const result = await this.baseService.delete(id);
       return ok(result);
     } catch (error) {
